Refactored and added, CustomUserDetailsService, SecurityConfig.java in progress errors., EventRestController almost done

+++ src/main/java/com/example/sso/repository/RegistrationRepository.java
@@ -4,6 +4,6 @@ import com.example.sso.model.Registration;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface RegistrationRepository extends JpaRepository<Registration, Integer> {
+public interface RegistrationRepository extends JpaRepository<Registration, Long> {
     Optional<Registration> findByEmail(String email);
 }
